bug#68507: [PATCH] doc: Fix example in list-transduce example.
While the `.' might be correct from a grammatical point of view (I do
not know), it turns the example into invalid scheme code, which is not
ideal.  New users (like me) might try to copy the whole line and wonder
why it does not work (like I did).  So delete it.

* doc/ref/srfi-modules.texi (SRFI-171 General Discussion): Delete the
trailing . from the example.

Signed-off-by: Ludovic Courtès <ludo@gnu.org>