Improves handling of searching for DLL foreign libraries
This patch improves dlopening DLLs.  First, the replacement for dlopen
in posix-w32.c was incorrect. LoadLibrary expects a wchar_t string that
is either an absolute path using backslashes or is just a filename.

Second, the foreign-library failed to handle searching for libraries
when no version number is given, for various types of DLLs.  If
a foreign library asks for "libfoo" on a DLL using system, Guile
needs to search for both "libfoo.dll" as well as "libfoo-X.dll" where
X is the major version number that libtool appends when it creates
DLLs.

Also, on Cygwin "libfoo" is "cygfoo.dll" and on MSYS, "libfoo"
is "msys-foo.dll", so search for those as well.

Lastly, it improves searching for DLLs which are usually in PATH
on DLL using systems rather than in LD_LIBRARY_PATH.

libguile/posix-w32.c (dlopen_w32): handle absolute paths of filenames only
module/system/foreign-library.scm (lib->msys): new helper function
(load-foreign-library): search PATH for DLLs, handle MSYS dlls